Move ExclusiveFileAccessManager to :files and remove duplicate from internal-integ-testing
