+++ platforms/core-runtime/wrapper-shared/src/main/java/org/gradle/wrapper/Install.java
@@ -16,6 +16,8 @@
 
 package org.gradle.wrapper;
 
+import org.gradle.internal.file.locking.ExclusiveFileAccessManager;
+
 import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.File;
