+++ subprojects/architecture-test/src/changes/archunit_store/internal-api-nullability.txt
@@ -5626,7 +5626,7 @@ Class <org.gradle.wrapper.Download$DefaultDownloadProgressListener> is not annot
 Class <org.gradle.wrapper.Download$ProxyAuthenticator> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Download.java:0)
 Class <org.gradle.wrapper.Download> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Download.java:0)
 Class <org.gradle.wrapper.DownloadProgressListener> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DownloadProgressListener.java:0)
-Class <org.gradle.wrapper.ExclusiveFileAccessManager> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ExclusiveFileAccessManager.java:0)
+Class <org.gradle.internal.file.locking.ExclusiveFileAccessManager> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ExclusiveFileAccessManager.java:0)
 Class <org.gradle.wrapper.GradleUserHomeLookup> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (GradleUserHomeLookup.java:0)
 Class <org.gradle.wrapper.GradleWrapperMain> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (GradleWrapperMain.java:0)
 Class <org.gradle.wrapper.IDownload> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (IDownload.java:0)
